feat(ota-image-tools): add lookup-image cmd (#18)
This PR adds a new lookup-image cmd in the ota-image-tools for looking up an OTA image payload from the OTA image with convenience.

An unique OTA image payload can be looked up with ecu-id and release-key, the command will print out the image-manifest of the looked up payload, or the image-config of the image if --image-config arg is specified.